backend/requirements.txt
# Web Framework ve Server
fastapi
uvicorn[standard]

# Ortam Değişkenleri Yönetimi
python-dotenv

# LangChain ve Google Yapay Zeka Entegrasyonu
langchain-core
langchain-groq
langgraph

# Opsiyonel: Veri doğrulama ve asenkron işlemler için genellikle gerekir
pydantic
httpx